taskdb_cal_{sync,export_all}: overwrite cal files only with newer modification
taskdb_cal_sync used to purge calendar directory in the beginning.
taskdb_cal_export_all used to dump files for each of the entries.

This seems to cause extraneous conflicts when series of frequent changes
to the same event are made on the side which is not the winner in
vdirsyncer's conflict resolution configuration.