ta: create tasks hour-aligned, hour-long
scheduled = CURRENT_TIMESTAMP is not so handy when editing with external
tools, as they leave odd numbers in seconds field, and, as a result,
events don't stack nicely no matter how hard you try.

Also creation of events with no positive duration resulted in some
events being invisible in Fastmail web UI. It is quite safe to assume
that each task requires some time, and 1h is a reasonable default.