schema: nop changes
Changes in results of schema dumping, not caused by anything particular
but the fact that for quite some time I have had non-mainline stuff in
my schema (functions and views) and have been submitting changes to
schema not completely but with `git add --patch`.

The advent of `alias` field obviates the need in the customizations i
had to use.